refactor: split utils.ts into multiple files
