+++ src/commands/slash/Main/blacklist/server.ts
@@ -1,4 +1,4 @@
-import { deleteConnections } from '#main/utils/ConnectedList.js';
+import { deleteConnections } from '#main/utils/ConnectedListUtils.js';
 import { emojis } from '#main/config/Constants.js';
 import { logBlacklist, logServerUnblacklist } from '#main/utils/HubLogger/ModLogs.js';
 import { t } from '#main/utils/Locale.js';
