+++ src/commands/slash/Main/connection/pause.ts
@@ -1,14 +1,15 @@
-import { updateConnection } from '#main/utils/ConnectedList.js';
+import { updateConnection } from '#main/utils/ConnectedListUtils.js';
 import { emojis } from '#main/config/Constants.js';
 import db from '#main/utils/Db.js';
 import { t } from '#main/utils/Locale.js';
-import { fetchCommands, findCommand, simpleEmbed } from '#main/utils/Utils.js';
+import { simpleEmbed } from '#main/utils/Utils.js';
 import {
   ChatInputCommandInteraction,
   channelMention,
   chatInputApplicationCommandMention as slashCmdMention,
 } from 'discord.js';
 import Connection from './index.js';
+import { fetchCommands, findCommand } from '#main/utils/CommandUtls.js';
 
 export default class Pause extends Connection {
   override async execute(interaction: ChatInputCommandInteraction): Promise<void> {
