+++ src/core/BaseClient.ts
@@ -5,10 +5,10 @@ import Scheduler from '#main/modules/SchedulerService.js';
 import ServerBlacklistManager from '#main/modules/ServerBlacklistManager.js';
 import UserDbManager from '#main/modules/UserDbManager.js';
 import type { RemoveMethods } from '#main/types/index.d.ts';
-import { isGuildTextBasedChannel } from '#main/utils/Channels.js';
-import { commandsMap, interactionsMap, loadCommandFiles } from '#main/utils/LoadCommands.js';
+import { isGuildTextBasedChannel } from '#main/utils/ChannelUtls.js';
+import { commandsMap, interactionsMap, loadCommandFiles } from '#main/utils/CommandUtls.js';
 import { loadLocales } from '#main/utils/Locale.js';
-import { RandomComponents } from '#main/utils/RandomComponents.js';
+import { RandomComponents } from '#main/RandomComponents.js';
 import { resolveEval } from '#main/utils/Utils.js';
 import { ClusterClient, getInfo } from 'discord-hybrid-sharding';
 import {
