+++ src/events/webhooksUpdate.ts
@@ -1,6 +1,6 @@
 import BaseEventListener from '#main/core/BaseEventListener.js';
-import { isGuildTextBasedChannel } from '#main/utils/Channels.js';
-import { updateConnection } from '#main/utils/ConnectedList.js';
+import { isGuildTextBasedChannel } from '#main/utils/ChannelUtls.js';
+import { updateConnection } from '#main/utils/ConnectedListUtils.js';
 import { emojis } from '#main/config/Constants.js';
 import db from '#main/utils/Db.js';
 import { t } from '#main/utils/Locale.js';
