+++ src/modules/BitFields.ts
