+++ src/tasks/storeMsgTimestamps.ts
@@ -1,5 +1,5 @@
 import cacheClient from '#main/utils/cache/cacheClient.js';
-import { updateConnection } from '#main/utils/ConnectedList.js';
+import { updateConnection } from '#main/utils/ConnectedListUtils.js';
 import { RedisKeys } from '#main/config/Constants.js';
 import Logger from '#main/utils/Logger.js';
 import { getAllDocuments, serializeCache } from '#main/utils/cache/cacheUtils.js';
