+++ src/utils/ConnectedListUtils.ts
