+++ src/utils/network/messageFormatters.ts
@@ -4,7 +4,7 @@ import type {
   EmbedFormatOpts,
 } from '#main/utils/network/Types.js';
 import Constants from '#main/config/Constants.js';
-import { censor } from '#main/utils/Profanity.js';
+import { censor } from '#main/utils/ProfanityUtils.js';
 import type { connectedList, hubs, userData } from '@prisma/client';
 import { EmbedBuilder, userMention, type WebhookMessageCreateOptions } from 'discord.js';
